Added new features to latex format.
New features:

- >= and <= are now treated as equalities.
- Add variable line length.
- Allows to format phantom environments that span multiple lines.
- Added eqnarray and gather to the equation environments.
- Added itemize as a text environment.
- Added reset option, this ignores the line splitting on the input
  text.
- Treat \pm and \mp equivalently to + and -.

Fixed bugs:

When closing a $ equation, the text afterwards was formatted as
equation, not as text.
Better treatment of begin/end blocks.
The tabs are now converted to 4 spaces. Before it using expandtabs
function, that does not the same (it was valid except for commented
lines)
Long comments now start at 2 comment tokens.
Correctly formatted multiline $ equations.
Don't split on \, and \; in equations.